Progress in the emulation of bzip2. However, 1. it doesn't work with arguments and 2. we need a thorough code review on MIPS
git-svn-id: svn+ssh://multi2sim.org/multi2sim/trunk@3032 a6503f54-2359-11de-8ce0-bfd7a313a9c3

+++ src/arch/mips/emu/ContextIsa.cc
@@ -344,7 +344,7 @@ void Context::ExecuteInst_LHU()
 	unsigned short int temp;
 	unsigned int addr = regs.getGPR(base) + misc::SignExtend32((signed)imm, 16);
 	memory->Read(addr, sizeof(unsigned short int), (char *)&temp);
-	regs.setGPR(rt, temp);
+	regs.setGPR(rt, (unsigned int)temp);
 }
